+++ kotlin-stdlib/common/main/kotlin/com/javiersc/kotlin/stdlib/Collections.kt
@@ -1,4 +1,4 @@
-@file:Suppress("MagicNumber", "TooManyFunctions")
+@file:Suppress("MagicNumber")
 
 package com.javiersc.kotlin.stdlib