+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `gradle -> 7.6`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.44`
 
 ## [0.1.0-alpha.12] - 2022-10-18