+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `easimon/wipe-cache -> v2`
 - `gradle -> 7.6.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.46`