+++ CHANGELOG.md
@@ -16,9 +16,9 @@
 
 ### Updated
 
+- `gradle -> 8.0.2`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.15`
 - `easimon/wipe-cache -> v2`
-- `gradle -> 7.6.1`
 
 ## [0.1.0-alpha.12] - 2022-10-18