+++ CHANGELOG.md
@@ -16,8 +16,8 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.16`
 - `gradle -> 8.0.2`
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.15`
 - `easimon/wipe-cache -> v2`
 
 ## [0.1.0-alpha.12] - 2022-10-18