+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.17`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.18`
 - `gradle -> 8.0.2`
 - `easimon/wipe-cache -> v2`