+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.6`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.8`
 - `gradle -> 8.0.2`
 - `easimon/wipe-cache -> v2`